+++ config/three-party-libs.js
@@ -1,7 +1,9 @@
 'use strict'
 
-module.exports = [
-  'core-js',
-  'rxjs',
-  'zone.js',
+// modules to separate from vendor module.
+const libs = [
+  'rxjs'
 ]
+
+// A regex to use in webpack configuartion
+module.exports = new RegExp(libs.join('|'))