+++ config/webpack-blocks/load-svg-sprite.js
@@ -1,0 +1,22 @@
+'use strict'
+
+const SpriteLoaderPlugin = require('svg-sprite-loader/plugin')
+
+/**
+ * Configuration of loader for svg sprites
+ */
+module.exports = (include, options = {}) => {
+  const plugins = options.extract ? [new SpriteLoaderPlugin()] : []
+
+  return {
+    module: {
+      rules: [{
+        test: /\.svg$/,
+        include,
+        loader: 'svg-sprite-loader',
+        options
+      }]
+    },
+    plugins
+  }
+}