+++ postcss.config.js
@@ -1,0 +1,10 @@
+'use strict'
+
+module.exports = {
+  plugins: {
+    'postcss-import': {},
+    'postcss-mixins': {},
+    'postcss-nested': {},
+    'postcss-cssnext': {}
+  }
+}