+++ src/app/app.component.ts
@@ -5,6 +5,6 @@ import { Component } from '@angular/core'
   styleUrls: ['./app.component.css'],
   template: `
     <h1>Hello {{ name }}</h1>
-  `,
+  `
 })
 export class AppComponent { name = 'Angular' }