+++ src/postcss.config.js
@@ -1,5 +1,0 @@
-'use strict'
-
-const postcssNiduscss = require('postcss-niduscss')
-
-module.exports = { ...postcssNiduscss }